+++ packages/svarog-ui-core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0](https://github.com/baaaaaaaaasowenyaaaaaaamamabeatsebaaah/svarog/compare/svarog-ui-core@2.0.0...svarog-ui-core@2.1.0) (2025-05-29)
+
+**Note:** Version bump only for package svarog-ui-core
+
+
+
+
+
 # [1.3.0](https://github.com/baaaaaaaaasowenyaaaaaaamamabeatsebaaah/svarog/compare/svarog-ui-core@1.2.4...svarog-ui-core@1.3.0) (2025-05-28)
 
 **Note:** Version bump only for package svarog-ui-core